web/requirements.txt
Flask==0.12.1
cassandra-driver==3.11.0
schedule==0.4.3
kafka-python==1.3.5